[Refactor] Remove all c style casts and other cleanups

+++ Arc/src/Arc/Core/UUID.h
@@ -22,7 +22,7 @@ namespace std
 	{
 		std::size_t operator()(const ArcEngine::UUID& uuid) const
 		{
-			return hash<uint64_t>()((uint64_t)uuid);
+			return hash<uint64_t>()(uuid);
 		}
 	};
 }
